pyproject.toml
[project]
name = "langchainexample"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "langchain>=0.3.9,<0.4.0",
    "langchain-community>=0.3.9,<0.4.0",
    "langchain-openai>=0.2.10,<0.3.0",
    "langchain-text-splitters>=0.3.2,<0.4.0",
    "openai>=1.57.4",
    "faiss-cpu>=1.8.0.post1",
    "tiktoken>=0.7.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.3.0",
    "html5lib>=1.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ingest-pycharm = "langchainexample.ingest_pycharm_docs:main"
ask-pycharm = "langchainexample.query_pycharm:main"

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
